backend/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.33.1
google-cloud-storage==2.19.0
SQLAlchemy==2.0.31
sqlalchemy-bigquery==1.11.0

grpc-google-iam-v1==0.13.1
grpcio==1.74.0
grpcio-status==1.71.2

langchain-pinecone==0.2.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7

protobuf==5.29.5

langchain
langchain_community
langchain_core
langchain_google_vertexai
langchain_openai
langchain_pinecone
langgraph
langserve
pybigquery
pydantic
uvicorn

anthropic
python-dotenv
Requests
typing_extensions
matplotlib
seaborn
numpy
pandas
Pillow
beautifulsoup4
fastapi
func_timeout
ipython
jupyter_client
bibtexparser
sse_starlette
tabulate
networkx
scipy
datasets